build-service/requirements.txt
# Build Service dependencies

# API framework
fastapi==0.115.11
uvicorn==0.29.0
pydantic==2.10.6
python-dotenv==1.0.1
python-multipart==0.0.9

# LangChain
langchain==0.3.21
langchain_community==0.3.20
langchain_core==0.3.46
langchain_neo4j==0.4.0
langchain_openai==0.3.9

# Neo4j
neo4j>=5.0.0
graphdatascience>=1.6

# Data
numpy==1.26.2
pandas==2.2.3

# Embeddings
sentence_transformers==4.1.0
faiss-cpu==1.9.0
tiktoken

# GCP auth (service-to-service on Cloud Run)
google-auth>=2.0.0
google-cloud-storage>=2.0.0
requests>=2.32.0

# Build pipeline utilities
rich>=13.0.0
psutil>=5.9.0
shutup>=0.2.0

# SEC parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0
